if role equals admin or editor the company selection is disabled and company is set to default value "Ticket Master". Also if you go back to the page "New User" the input fields are empty.
