+++ src/main/java/org/dhbw/mosbach/ai/tickets/beans/UserBean.java
@@ -4,6 +4,7 @@ import com.google.common.collect.ImmutableList;
 import org.dhbw.mosbach.ai.tickets.database.UserDAO;
 import org.dhbw.mosbach.ai.tickets.model.Role;
 import org.dhbw.mosbach.ai.tickets.model.User;
+import org.dhbw.mosbach.ai.tickets.view.AdminView;
 
 import javax.enterprise.context.SessionScoped;
 import javax.faces.application.FacesMessage;
@@ -24,17 +25,13 @@ public class    UserBean extends AbstractBean {
     @Inject
     private SecurityBean securityBean;
 
+    @Inject
+    private AdminView adminView;
+
     private User currentUser;
 
     private List<User> searchResult;
     private String searchString = "";
-
-    private String login_id;
-    private String name;
-    private String companyName;
-    private String email;
-    private String password;
-    private String role;
 
     private List<Role> roles;
 
@@ -52,6 +49,12 @@ public class    UserBean extends AbstractBean {
         userDAO.changePassword(user, password);
         saveUser(user);
 
+        adminView.setName("");
+        adminView.setLogin_id("");
+        adminView.setEmail("");
+        adminView.setRole("customer");
+        adminView.setCompanyName("");
+        adminView.setPassword("");
         return VIEW_USERS;
     }
 
@@ -122,6 +125,7 @@ public class    UserBean extends AbstractBean {
     public void fetchAllCompanies() {
         companies = new ArrayList<>();
         companies = userDAO.getCompanies();
+        companies.remove("Ticket Master");
     }
 
     public List<User> getEditors() {
@@ -145,53 +149,6 @@ public class    UserBean extends AbstractBean {
         this.currentUser = currentUser;
     }
 
-    public String getLogin_id() {
-        return login_id;
-    }
-
-    public void setLogin_id(String login_id) {
-        this.login_id = login_id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getCompanyName() {
-        return companyName;
-    }
-
-    public void setCompanyName(String companyName) {
-        this.companyName = companyName;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public String getRole() {
-        return role;
-    }
-
-    public void setRole(String role) {
-        this.role = role;
-    }
 
     public List<String> getCompanies() {
